move util functions to new file
