+++ src/hook/hook.go
@@ -113,28 +113,3 @@ func (m *DNSWebhook) addOrUpdateDNSRecord(w http.ResponseWriter, r *http.Request
 	w.WriteHeader(http.StatusNoContent)
 	return nil
 }
-
-// write200Response writes the response to be sent
-func writeJSONResponse(payload interface{}, statusCode int, w http.ResponseWriter) {
-	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(statusCode)
-
-	if payload != nil {
-		types.PanicIfError(json.NewEncoder(w).Encode(payload))
-	}
-
-	logrus.Infof("%d Response sent. Payload: %v", statusCode, payload)
-}
-
-// handleError recovers from a panic
-func handleError(w http.ResponseWriter) {
-	r := recover()
-	if r != nil {
-		err := types.InternalServerError("An internal server error occurred, please contact the system administrator.", nil)
-		if e, ok := r.(*types.Error); ok {
-			err = e
-		}
-		logrus.Error(err)
-		writeJSONResponse(err, err.Code, w)
-	}
-}
